Move grid to separate component, check bounds, only letters
